refactor: better transaction subscriber
introduces a concept of ITransactionProcess which handles the logic of processing a transaction, and does nothing else. allows the queueing logic to be abstracted away and code coverage the be better
